Move output related files.
This will make adding a trait for isFresh less smelly.
